Оптимизирует код, рефакторинг компонента теймеров, изменение архитектуры
