+++ src/auto-imports.d.ts
@@ -10,7 +10,6 @@ declare global {
   const SideBar: typeof import('./components/index')['SideBar']
   const Switcher: typeof import('./components/index')['Switcher']
   const TimeModal: typeof import('./components/index')['TimeModal']
-  const Timer: typeof import('./components/index')['Timer']
   const createRef: typeof import('react')['createRef']
   const forwardRef: typeof import('react')['forwardRef']
   const lazy: typeof import('react')['lazy']
