+++ src/components/index.ts
@@ -2,7 +2,6 @@ import TimeModal from "./time-modal/TimeModal";
 import SideBar from "./side-bar/SideBar";
 import ModalContainer from "./modal-container/ModalContainer";
 import Counter from "./counter/Counter";
-import Timer from './timer/Timer'
 import Pointer from "./pointer/Pointer";
 import Button from './button/Button'
 import Switcher from './switcher/Switcher'
@@ -14,7 +13,6 @@ export {
 	SideBar,
 	ModalContainer,
 	Counter,
-	Timer,
 	Pointer,
 	Button,
 	Switcher,
