+++ src/pages/main/components/timers/Timer.module.scss
