+++ src/pages/main/hooks/useTimer.tsx
@@ -1,3 +1,4 @@
+
 import useSound from "use-sound";
 
 import sound from "@/assets/sounds/sound.mp3";
